requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.5.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.42.0
django-heroku==0.3.1
django-picklefield==3.0.1
ffmpeg==1.4
filelock==3.0.12
fuzzywuzzy==0.18.0
gunicorn==20.0.4
icalendar==4.0.7
idna==2.10
isort==5.5.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
multidict==4.7.6
networkx==2.4
numpy==1.18.2
oauthlib==3.1.0
opuslib==3.0.1
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.5.0
Pygments==2.7.2
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rich==9.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tqdm==4.51.0
tweepy==3.9.0
typing-extensions==3.7.4.3
urllib3==1.25.10
validators==0.18.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.11.12
